Autoclick on favorite panels
Added button in the favorite menu to choose an autoclick option for the
chosen fav. Let the user choose a timer to set it, and choose what
button to click if there are multiples
Added a global hook on redraw to ensure inserted elements are still
there.
Needs more testing, and needs to add persist behavior.